Changed Customer to CustomerModel, moved Screen from originally being in CustomerDatabase, to CustomerListScreen Changed Method for Removeproduct into RemoveProduct instead of RemoveCompany made CustomerModel able to take nulls begun Working on CustomerDetailScreen and CustomerListScreen
