+++ H1ERPSystem2023/Databasefiles/CompanyDatabase.cs
@@ -19,7 +19,6 @@ namespace H1ERPSystem2023.Databasefiles
         {
             _addCompanies();
         }
-
         //GetCompany Gets an ID i program.cs (by the user), and uses that with the foreach to take all companies
         //and check whichever one has a matching ID, so it can return the information.
         public CompanyModel GetCompany(int ID)
