+++ H1ERPSystem2023/Domain Model/CustomerModel.cs
@@ -8,7 +8,7 @@
         /// <summary>
         /// Creates a new Customer With a Person Parameters
         /// </summary>
-        public CustomerModel(int personId, string firstName, string lastName, AddressModel address, string phoneNumber, string emailAddress, string customerNumber, DateTime lastPurchaseDate) : base(personId, firstName, lastName, address, phoneNumber, emailAddress)
+        public CustomerModel(int personId, string firstName, string lastName, AddressModel? address, string phoneNumber, string emailAddress, string customerNumber, DateTime? lastPurchaseDate) : base(personId, firstName, lastName, address, phoneNumber, emailAddress)
         {
             #region Person
             this.PersonId = personId;
