+++ H1ERPSystem2023/Screens/CustomerDetailScreen.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace H1_ERP_System_2023.Screens
+{
+    internal class CustomerDetailScreen
+    {
+    }
+}
